added more checks to ensure data integrity

+++ src/main/java/org/projectmanagement/application/services/RolesPermissionsServiceImpl.java
@@ -4,17 +4,20 @@ package org.projectmanagement.application.services;
 import jakarta.transaction.Transactional;
 import org.projectmanagement.application.dto.roles_permissions.RolesPermissionsCreate;
 import org.projectmanagement.application.dto.roles_permissions.RolesPermissionsUpdate;
+import org.projectmanagement.domain.entities.Permissions;
 import org.projectmanagement.domain.entities.RolesPermissions;
 import org.projectmanagement.domain.entities.Roles;
 import org.projectmanagement.domain.exceptions.ResourceAlreadyExistsException;
 import org.projectmanagement.domain.exceptions.ResourceNotFoundException;
 import org.projectmanagement.domain.exceptions.SystemRoleUpdateException;
+import org.projectmanagement.domain.repository.PermissionsRepository;
 import org.projectmanagement.domain.repository.RolesPermissionsRepository;
 import org.projectmanagement.domain.repository.RolesRepository;
 import org.projectmanagement.domain.services.RolesPermissionsService;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 import java.util.UUID;
 
@@ -22,12 +25,15 @@ import java.util.UUID;
 public class RolesPermissionsServiceImpl implements RolesPermissionsService {
     private final RolesPermissionsRepository rolesPermissionsRepository;
     private final RolesRepository rolesRepository;
+    private final PermissionsRepository permissionsRepository;
 
     RolesPermissionsServiceImpl(RolesPermissionsRepository rolesPermissionsRepository
             , RolesRepository rolesRepository
+            , PermissionsRepository permissionsRepository
     ) {
         this.rolesPermissionsRepository = rolesPermissionsRepository;
         this.rolesRepository = rolesRepository;
+        this.permissionsRepository = permissionsRepository;
     }
 
 
@@ -59,7 +65,7 @@ public class RolesPermissionsServiceImpl implements RolesPermissionsService {
         Roles roleFromDB = rolesRepository.findByExactName(dto.roleName(), companyId).orElse(null);
 
         if (roleFromDB != null) {
-            throw new ResourceAlreadyExistsException("Roles with name: " + dto.roleName() + "already exists.");
+            throw new ResourceAlreadyExistsException("Role with name: " + dto.roleName() + " already exists.");
        }
         Roles createdRole = rolesRepository.save(Roles.builder()
                         .name(dto.roleName())
@@ -67,6 +73,12 @@ public class RolesPermissionsServiceImpl implements RolesPermissionsService {
                         .isSystemRole(false)
                         .isDeleted(false)
                         .build());
+
+        List<UUID> permissions = permissionsRepository.findAllPermissionIds();
+
+        if (!new HashSet<>(permissions).containsAll(dto.permissions())) {
+           throw new ResourceNotFoundException("One of the given Permissions was not found.");
+        }
 
         for (UUID permissionId : dto.permissions()) {
             rolesPermissions.add(
@@ -91,13 +103,20 @@ public class RolesPermissionsServiceImpl implements RolesPermissionsService {
             throw new SystemRoleUpdateException("System Roles can not be updated. The role: " + roleFromDB.getName() + " is a system role.");
         }
 
-        List<UUID> permissionsOfRole = rolesPermissionsRepository.findAllPermissionsOfRoleByRoleId(roleId);
+        List<UUID> permissions = permissionsRepository.findAllPermissionIds();
+        List<UUID> rolesPermissionsFromDb = rolesPermissionsRepository.findAllPermissionsOfRoleByRoleId(roleId);
+
+        if (!new HashSet<>(permissions).containsAll(dto.permissions())) {
+            throw new ResourceNotFoundException("One of the given Permissions was not found.");
+        }
 
         for (UUID permissionId : dto.permissions()) {
-            if (!permissionsOfRole.contains(permissionId)) {
-                rolesPermissions.add(
-                        rolesPermissionsRepository.save(new RolesPermissions(UUID.randomUUID(), roleId, permissionId))
-                );
+            //controls that duplicates are not created
+            if(!rolesPermissionsFromDb.contains(permissionId)) {
+                rolesPermissions.add(rolesPermissionsRepository.save(RolesPermissions.builder()
+                        .roleId(roleId)
+                        .permissionId(permissionId)
+                        .build()));
             }
         }
 
@@ -117,6 +136,11 @@ public class RolesPermissionsServiceImpl implements RolesPermissionsService {
             throw new SystemRoleUpdateException("System Roles can not be updated. The role: " + roleFromDB.getName() + " is a system role.");
         }
 
+        List<UUID> permissions = permissionsRepository.findAllPermissionIds();
+
+        if (!new HashSet<>(permissions).containsAll(dto.permissions())) {
+            throw new ResourceNotFoundException("One of the given Permissions was not found.");
+        }
 
         for (UUID permissionId : dto.permissions()) {
             rolesPermissionsRepository.deleteById(permissionId);
